requirements.txt
fastapi[standard]
uvicorn[standard]
torch
torchvision
pillow
pytest
flake8

